# improve fp16 performance of slice_grad