… paddle/fluid/operators/slice_op.cu