### PR types
<!-- One of [ New features | Bug fixes | Function optimization | Performance optimization | Breaking changes | Others ] --> Performance optimization

### PR changes
<!-- One of [ OPs | APIs | Docs | Others ] --> OPs

### Describe
<!-- Describe what this PR does --> improve fp16 performance of slice_grad

slice_grad的fp16特化kernel性能较差，因此本PR移除了特化kernel。性能数据如下：
- 测试环境：v100, cuda10
- 以下输入均为fp16，对比total gpu time

**GPU FP16 performance**
|op|input shape|start|end|axes|before|after| speed up |
|---|---|---|---|---|---|---|---|
|slice_grad|[16L, 800L]|[200]|[400]|[1]| 0.012141 ms|0.005447 ms|2.2x| 
|slice_grad|[35L, 16L, 1500L]|[34]|[35]|[0]|0.276627ms|0.198814 ms|1.4x|  
|slice_grad|[2L, 16L, 1500L]|[0]|[1]|[0]|0.012747 ms|0.006055 ms|2.1x|